+++ src/Candoumbe.Pipelines/Components/Configuration.cs
@@ -1,27 +1,24 @@
-﻿using JetBrains.Annotations;
-
-using Nuke.Common.Tooling;
+﻿using Nuke.Common.Tooling;
 
 using System.ComponentModel;
 
 namespace Candoumbe.Pipelines.Components
 {
     /// <summary>
-    /// Configuration that can be used throughout CI/CD pipelines
+    /// Configuration that can be used to compile an application
     /// </summary>
-    [PublicAPI]
     [TypeConverter(typeof(TypeConverter<Configuration>))]
     public class Configuration : Enumeration
     {
         /// <summary>
-        /// The "Debug" mode
+        /// Debug mode
         /// </summary>
-        public static Configuration Debug => new() { Value = nameof(Debug) };
+        public static readonly Configuration Debug = new() { Value = nameof(Debug) };
 
         /// <summary>
-        /// The "release" mode
+        /// Release mode
         /// </summary>
-        public static Configuration Release => new() { Value = nameof(Release) };
+        public static readonly Configuration Release = new() { Value = nameof(Release) };
 
         ///<inheritdoc/>
         public static implicit operator string(Configuration configuration) => configuration.Value;